RAG_systems/Adaptive_RAG/requirements.txt
openai>=1.0.0
langchain>=0.1.0
faiss-cpu>=1.7.0
neo4j>=5.0.0
python-dotenv>=1.0.0
tqdm>=4.65.0
pydantic>=2.0.0
numpy>=1.24.0
streamlit>=1.28.0
